Changing Rtv to fit GBuffer

+++ RasterizerDemo/Src/FileReader/Parsers/IFileParser.hpp
@@ -34,8 +34,8 @@ protected:
         // Gets line from file
         while(std::getline(inFile, line))
         {
-            int start = 0;
-            int end = 0;
+            size_t start = 0;
+            size_t end = 0;
             
             std::vector<std::string> tokenVector;
